pyproject.toml
[tool.poetry]
name = "dro-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "dro_project"}]

[tool.poetry.dependencies]
python = "^3.11"
pyserial = "^3.5"
matplotlib = "^3.6.2"
tk = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
